first tester errors were checked up to strrchr

+++ ft_bzero.c
@@ -19,7 +19,7 @@ void	ft_bzero(void *s, size_t n)
 
 {
 	char	*v;
-	int		b;
+	size_t	b;
 
 	v = (char *)s;
 	b = 0;
